change how we keep time so we can easily sync with an external host
instead of counting samples we now keep a fractional beat counter, where a beat is two steps in the sequencer. the beat counter ranges between 0 and 1 so that we can check its value against our [0,1] shuffle parameter in order to play our up-beats at the correct time. the beat counter is incremented every audio sample at a rate determined by Tempo and Step Length. this means that we can respond to tempo changes in a much smoother way. this also enables us to use the Host's PPQ position when our Clock Source is set to External in order to *snap* the beat counter to the correct value at the beginning of every Process call, which is important for staying in sync with the Host's location when entering play mode mid-beat.